original/showcases/embeddings-service/requirements.txt
torch==2.1.0
sentence-transformers==2.2.2
transformers==4.35.0
pillow==10.1.0
ftfy==6.1.1
regex==2023.10.3
numpy==1.24.3
scikit-learn==1.3.2
